Tokenizer almost finished
I discovered this did not go through to Himani's Repo... trying again

This is the partially completed Tokenizer Input portion of the first set of outlines. There are two sections of suedo code that still need to be replaced in the next() of the Tokienzer.java class and the constructor of the same file I am having a little trouble with. Everything else should be there :) and I am working to figure out these issues as we speak :)
Learning more about Github everyday :P