Fixed bugs
There are still some errors that I can tell. From what I am seeing, The
Tokenizer was implented in Chat.java . When we made new commands with
the @overrride command, it used the old system of parsing the words. I
need to figure out how to change this: final String message =
args.hasNext() ? args.nextLine().trim() : ""; to what it needs to be?
This looks like the final step, but i got it to clean and build (with
errors, but soon to be fixed, or fixed tomorrow)